requirements.txt
pandas
python-dateutil
pytz
six
tzdata
python-dotenv
tqdm
scikit-learn
telethon

# --- NER & ML stack ---
numpy>=1.23
matplotlib>=3.5
torch>=2.2
transformers>=4.40
datasets>=3.5
evaluate>=0.4

# Interpretability
shap
lime

# Data & model versioning
# Includes Google Drive support for DVC remotes
dvc[gdrive]